feat: support `Evaluates:` group
