+++ src/main/kotlin/mathlingua/Main.kt
@@ -33,7 +33,7 @@ import mathlingua.common.support.ParseError
 import mathlingua.common.support.ValidationFailure
 import mathlingua.common.support.ValidationSuccess
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.defines.DefinesGroup
-import mathlingua.common.chalktalk.phase2.ast.group.toplevel.states.StatesGroup
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.states.StatesGroup
 import java.io.File
 import java.nio.file.Paths
 import kotlin.system.exitProcess
