+++ src/main/kotlin/mathlingua/common/chalktalk/phase2/ast/group/clause/expands/AsSection.kt
@@ -18,6 +18,7 @@ package mathlingua.common.chalktalk.phase2.ast.group.clause.expands
 
 import mathlingua.common.support.MutableLocationTracker
 import mathlingua.common.chalktalk.phase1.ast.Phase1Node
+import mathlingua.common.chalktalk.phase1.ast.Section
 import mathlingua.common.chalktalk.phase2.ast.clause.ClauseListNode
 import mathlingua.common.chalktalk.phase2.CodeWriter
 import mathlingua.common.chalktalk.phase2.ast.common.Phase2Node
@@ -44,6 +45,8 @@ data class AsSection(val clauses: ClauseListNode) : Phase2Node {
             )
             )
 }
+
+fun isAsSection(sec: Section) = sec.name.text == "as"
 
 fun validateAsSection(node: Phase1Node, tracker: MutableLocationTracker) = validateClauseList(
         AtLeast(1),
