+++ src/main/kotlin/mathlingua/common/chalktalk/phase2/ast/group/toplevel/defineslike/states/StatesGroup.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package mathlingua.common.chalktalk.phase2.ast.group.toplevel.views
+package mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.states
 
 import mathlingua.common.support.MutableLocationTracker
 import mathlingua.common.support.ParseError
@@ -37,33 +37,41 @@ import mathlingua.common.chalktalk.phase2.ast.section.*
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.metadata.section.MetaDataSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.metadata.section.validateMetaDataSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.TopLevelGroup
-import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.foundation.DefinesStatesOrViews
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.UsingSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.WhenSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.WrittenSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.foundation.DefinesStatesOrViews
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.validateUsingSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.validateWhenSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.validateWrittenSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.topLevelToCode
 import mathlingua.common.transform.signature
 import mathlingua.common.support.validationFailure
 import mathlingua.common.support.validationSuccess
 
-data class ViewsGroup(
+data class StatesGroup(
     val signature: String?,
     val id: IdStatement,
-    val viewsSection: ViewsSection,
-    val singleFromSection: SingleFromSection,
-    val singleToSection: SingleToSection,
-    val asSection: SingleAsSection,
+    val statesSection: StatesSection,
+    val whenSection: WhenSection?,
+    val thatSections: List<ThatSection>,
     val usingSection: UsingSection?,
+    val writtenSection: WrittenSection?,
     override val metaDataSection: MetaDataSection?
 ) : TopLevelGroup(metaDataSection), DefinesStatesOrViews {
 
     override fun forEach(fn: (node: Phase2Node) -> Unit) {
         fn(id)
-        fn(viewsSection)
-        fn(singleFromSection)
-        fn(singleToSection)
-        fn(asSection)
+        fn(statesSection)
+        if (whenSection != null) {
+            fn(whenSection)
+        }
+        thatSections.forEach(fn)
         if (usingSection != null) {
             fn(usingSection)
+        }
+        if (writtenSection != null) {
+            fn(writtenSection)
         }
         if (metaDataSection != null) {
             fn(metaDataSection)
@@ -71,37 +79,37 @@ data class ViewsGroup(
     }
 
     override fun toCode(isArg: Boolean, indent: Int, writer: CodeWriter): CodeWriter {
+        val sections = mutableListOf(statesSection, whenSection)
+        sections.addAll(thatSections)
+        sections.add(usingSection)
+        sections.add(writtenSection)
+        sections.add(metaDataSection)
         return topLevelToCode(
-            writer,
-            isArg,
-            indent,
-            id,
-            viewsSection,
-            singleFromSection,
-            singleToSection,
-            asSection,
-            usingSection,
-            metaDataSection
+                writer,
+                isArg,
+                indent,
+                id,
+                *sections.toTypedArray()
         )
     }
 
     override fun transform(chalkTransformer: (node: Phase2Node) -> Phase2Node) = chalkTransformer(
-        ViewsGroup(
-        signature = signature,
-        id = id.transform(chalkTransformer) as IdStatement,
-        viewsSection = viewsSection.transform(chalkTransformer) as ViewsSection,
-        singleFromSection = singleFromSection.transform(chalkTransformer) as SingleFromSection,
-        singleToSection = singleToSection.transform(chalkTransformer) as SingleToSection,
-        asSection = asSection.transform(chalkTransformer) as SingleAsSection,
-        usingSection = usingSection?.transform(chalkTransformer) as UsingSection?,
-        metaDataSection = metaDataSection?.transform(chalkTransformer) as MetaDataSection?
+        StatesGroup(
+            signature = signature,
+            id = id.transform(chalkTransformer) as IdStatement,
+            statesSection = statesSection.transform(chalkTransformer) as StatesSection,
+            whenSection = whenSection?.transform(chalkTransformer) as WhenSection?,
+            thatSections = thatSections.map { chalkTransformer(it) as ThatSection },
+            usingSection = usingSection?.transform(chalkTransformer) as UsingSection?,
+            writtenSection = writtenSection?.transform(chalkTransformer) as WrittenSection?,
+            metaDataSection = metaDataSection?.transform(chalkTransformer) as MetaDataSection?
     )
     )
 }
 
-fun isViewsGroup(node: Phase1Node) = firstSectionMatchesName(node, "Views")
+fun isStatesGroup(node: Phase1Node) = firstSectionMatchesName(node, "States")
 
-fun validateViewsGroup(groupNode: Group, tracker: MutableLocationTracker): Validation<ViewsGroup> {
+fun validateStatesGroup(groupNode: Group, tracker: MutableLocationTracker): Validation<StatesGroup> {
     val errors = ArrayList<ParseError>()
     val group = groupNode.resolve()
     var id: IdStatement? = null
@@ -138,57 +146,63 @@ fun validateViewsGroup(groupNode: Group, tracker: MutableLocationTracker): Valid
     try {
         sectionMap = identifySections(
             sections,
-            "Views", "from", "to", "as", "using?", "Metadata?"
+            "States", "when?", "that", "using?", "where?", "written?", "Metadata?"
         )
     } catch (e: ParseError) {
         errors.add(ParseError(e.message, e.row, e.column))
         return validationFailure(errors)
     }
 
-    val viewsNode = sectionMap["Views"]!!
-    val fromNode = sectionMap["from"]!!
-    val toNode = sectionMap["to"]!!
-    val asNode = sectionMap["as"]!!
-    val usingNode = sectionMap["using"]
-    val metadataNode = sectionMap["Metadata"]
+    val definesLike = sectionMap["States"]!!
+    val whenNode = sectionMap["when"] ?: emptyList()
+    val ends = sectionMap["that"]!!
+    val using = sectionMap["using"] ?: emptyList()
+    val written = sectionMap["written"] ?: emptyList()
+    val metadata = sectionMap["Metadata"] ?: emptyList()
 
-    var viewsSection: ViewsSection? = null
-    when (val validation = validateViewsSection(viewsNode[0], tracker)) {
-        is ValidationSuccess -> viewsSection = validation.value
-        is ValidationFailure -> errors.addAll(validation.errors)
+    var statesSection: StatesSection? = null
+    when (val representsValidation = validateStatesSection(definesLike[0], tracker)) {
+        is ValidationSuccess -> statesSection = representsValidation.value
+        is ValidationFailure -> errors.addAll(representsValidation.errors)
     }
 
-    var singleFromSection: SingleFromSection? = null
-    when (val validation = validateSingleFromSection(fromNode[0], tracker)) {
-        is ValidationSuccess -> singleFromSection = validation.value
-        is ValidationFailure -> errors.addAll(validation.errors)
+    var whenSection: WhenSection? = null
+    if (whenNode.isNotEmpty()) {
+        when (val assumingValidation = validateWhenSection(whenNode[0], tracker)) {
+            is ValidationSuccess -> whenSection = assumingValidation.value
+            is ValidationFailure -> errors.addAll(assumingValidation.errors)
+        }
     }
 
-    var singleToSection: SingleToSection? = null
-    when (val validation = validateSingleToSection(toNode[0], tracker)) {
-        is ValidationSuccess -> singleToSection = validation.value
-        is ValidationFailure -> errors.addAll(validation.errors)
-    }
-
-    var singleAsSection: SingleAsSection? = null
-    when (val validation = validateSingleAsSection(asNode[0], tracker)) {
-        is ValidationSuccess -> singleAsSection = validation.value
-        is ValidationFailure -> errors.addAll(validation.errors)
+    val endSections = mutableListOf<ThatSection>()
+    for (end in ends) {
+        when (val endValidation = validateThatSection(end, tracker)) {
+            is ValidationSuccess -> endSections.add(endValidation.value)
+            is ValidationFailure -> errors.addAll(endValidation.errors)
+        }
     }
 
     var usingSection: UsingSection? = null
-    if (usingNode != null) {
-        when (val validation = validateUsingSection(usingNode[0], tracker)) {
-            is ValidationSuccess -> usingSection = validation.value
-            is ValidationFailure -> errors.addAll(validation.errors)
+    if (using.isNotEmpty()) {
+        when (val aliasValidation = validateUsingSection(using[0], tracker)) {
+            is ValidationSuccess -> usingSection = aliasValidation.value
+            is ValidationFailure -> errors.addAll(aliasValidation.errors)
+        }
+    }
+
+    var writtenSection: WrittenSection? = null
+    if (written.isNotEmpty()) {
+        when (val writtenValidation = validateWrittenSection(written[0], tracker)) {
+            is ValidationSuccess -> writtenSection = writtenValidation.value
+            is ValidationFailure -> errors.addAll(writtenValidation.errors)
         }
     }
 
     var metaDataSection: MetaDataSection? = null
-    if (metadataNode != null) {
-        when (val validation = validateMetaDataSection(metadataNode[0], tracker)) {
-            is ValidationSuccess -> metaDataSection = validation.value
-            is ValidationFailure -> errors.addAll(validation.errors)
+    if (metadata.isNotEmpty()) {
+        when (val metaDataValidation = validateMetaDataSection(metadata[0], tracker)) {
+            is ValidationSuccess -> metaDataSection = metaDataValidation.value
+            is ValidationFailure -> errors.addAll(metaDataValidation.errors)
         }
     }
 
@@ -197,14 +211,15 @@ fun validateViewsGroup(groupNode: Group, tracker: MutableLocationTracker): Valid
     } else validationSuccess(
         tracker,
         groupNode,
-        ViewsGroup(
+        StatesGroup(
             id?.signature(),
-            id!!,
-            viewsSection!!,
-            singleFromSection!!,
-            singleToSection!!,
-            singleAsSection!!,
+            id!!, statesSection!!,
+            whenSection,
+            // the end sections are in reverse order so they
+            // must be reversed here to be in the correct order
+            endSections.reversed(),
             usingSection,
+            writtenSection,
             metaDataSection
         )
     )
