+++ src/main/kotlin/mathlingua/common/chalktalk/phase2/ast/group/toplevel/defineslike/states/StatesSection.kt
@@ -1,0 +1,78 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.states
+
+import mathlingua.common.chalktalk.phase1.ast.Phase1Node
+import mathlingua.common.chalktalk.phase1.ast.Section
+import mathlingua.common.chalktalk.phase1.ast.getColumn
+import mathlingua.common.chalktalk.phase1.ast.getRow
+import mathlingua.common.chalktalk.phase2.CodeWriter
+import mathlingua.common.chalktalk.phase2.ast.common.Phase2Node
+import mathlingua.common.support.MutableLocationTracker
+import mathlingua.common.support.ParseError
+import mathlingua.common.support.Validation
+import mathlingua.common.support.validationFailure
+import mathlingua.common.support.validationSuccess
+
+class StatesSection : Phase2Node {
+    override fun forEach(fn: (node: Phase2Node) -> Unit) {}
+
+    override fun toCode(isArg: Boolean, indent: Int, writer: CodeWriter): CodeWriter {
+        writer.writeIndent(isArg, indent)
+        writer.writeHeader("States")
+        return writer
+    }
+
+    override fun transform(chalkTransformer: (node: Phase2Node) -> Phase2Node) = chalkTransformer(this)
+}
+
+fun validateStatesSection(node: Phase1Node, tracker: MutableLocationTracker): Validation<StatesSection> {
+    val errors = ArrayList<ParseError>()
+    if (node !is Section) {
+        errors.add(
+                ParseError(
+                        "Expected a StatesSection",
+                        getRow(node), getColumn(node)
+                )
+        )
+    }
+
+    val sect = node as Section
+    if (sect.args.isNotEmpty()) {
+        errors.add(
+                ParseError(
+                        "A States cannot have any arguments",
+                        getRow(node), getColumn(node)
+                )
+        )
+    }
+
+    if (sect.name.text != "States") {
+        errors.add(
+                ParseError(
+                        "Expected a section named States",
+                        getRow(node), getColumn(node)
+                )
+        )
+    }
+
+    return if (errors.isNotEmpty()) {
+        validationFailure(errors)
+    } else {
+        validationSuccess(tracker, node, StatesSection())
+    }
+}
