+++ src/main/kotlin/mathlingua/common/translate/LatexTranslator.kt
@@ -49,18 +49,18 @@ import mathlingua.common.chalktalk.phase2.ast.group.clause.iff.IffSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.defines.MeansSection
 import mathlingua.common.chalktalk.phase2.ast.group.clause.not.NotSection
 import mathlingua.common.chalktalk.phase2.ast.group.clause.or.OrSection
-import mathlingua.common.chalktalk.phase2.ast.group.toplevel.states.StatesSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.states.StatesSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.resource.ResourceSection
 import mathlingua.common.chalktalk.phase2.ast.group.clause.exists.SuchThatSection
 import mathlingua.common.chalktalk.phase2.ast.section.TextSection
-import mathlingua.common.chalktalk.phase2.ast.group.toplevel.states.ThatSection
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.states.ThatSection
 import mathlingua.common.chalktalk.phase2.ast.group.clause.`if`.ThenSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.resultlike.theorem.TheoremSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.shared.WhereSection
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.resultlike.axiom.AxiomGroup
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.resultlike.conjecture.ConjectureGroup
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.defines.DefinesGroup
-import mathlingua.common.chalktalk.phase2.ast.group.toplevel.states.StatesGroup
+import mathlingua.common.chalktalk.phase2.ast.group.toplevel.defineslike.states.StatesGroup
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.resource.ResourceGroup
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.resultlike.theorem.TheoremGroup
 import mathlingua.common.chalktalk.phase2.ast.group.toplevel.TopLevelGroup
